jobs/us_regions/requirements.txt
python-dotenv
pandas
requests
SQLAlchemy
psycopg2-binary
us
geopy
